requirements.txt
crewai>=0.11.0
python-telegram-bot>=20.7
supabase>=2.3.1
python-dotenv>=1.0.0
openai>=1.12.0
langchain>=0.3.0
tiktoken>=0.5.2
python-docx>=1.0.1
PyPDF2>=3.0.0
numpy>=1.24.0
tenacity>=8.2.0 